fixed organization (grouped user related functions in userController.js, userRoute.js)
